--- lib/types/token.ts
@@ -1,7 +1,0 @@
-export interface TokenResponse {
-  access_token: string
-  token_type: string
-  expires_in: number
-  scope: string
-  error?: string
-}
